+++ docs/docs/.vuepress/components/SceneColorBg.vue
@@ -1,0 +1,13 @@
+<template>
+  <tv-scene class="scene" clear-color="#bcedce"></tv-scene>
+</template>
+
+<script lang="ts" setup></script>
+
+<style>
+.scene {
+  margin-top: 10px;
+  width: 100%;
+  height: 300px;
+}
+</style>