+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.3] - 2025-02-12
+
+### Add
+
+- Add components prefix.
+- Add scene clear color.
+
+---
+
 ## [0.0.2] - 2025-02-09
 
 ### Modified